+++ unit_tests/actuator/CMakeLists.txt
@@ -3,6 +3,7 @@ target_sources(${utest_ex_name} PRIVATE
    ${CMAKE_CURRENT_SOURCE_DIR}/UnitTestActuatorBulk.C
    ${CMAKE_CURRENT_SOURCE_DIR}/UnitTestActuatorNGP.C
    ${CMAKE_CURRENT_SOURCE_DIR}/UnitTestActuatorSearch.C
+   ${CMAKE_CURRENT_SOURCE_DIR}/UnitTestActuatorFunctors.C
 )
 if(ENABLE_OPENFAST)
    target_sources(${utest_ex_name} PRIVATE
